deleted drivers and operators table, added and created members table, added forign key in vans table

+++ arkila/database/migrations/2018_01_24_104626_create_destinations_table.php
@@ -17,7 +17,7 @@ class CreateDestinationsTable extends Migration
             $table->increments('destination_id');
             $table->string('description');
             $table->enum('terminal', ['Cabanatuan City', 'San Jose City']);
-            $table->double('amount')
+            $table->decimal('amount', 4, 2)
             ->unsigned();
             $table->timestamps();
